added challenge contrlller module, added custom mode2 popup module #495
