+++ src/js/global-dependencies.js
@@ -53,3 +53,5 @@ import * as LanguagePicker from "./language-picker";
 import * as UI from "./ui";
 import * as Commandline from "./commandline";
 import * as CommandlineLists from "./commandline-lists";
+import * as ChallengeController from "./challenge-controller";
+import * as CustomMode2Popup from "./custom-mode2-popup";
