+++ src/js/script.js
@@ -401,101 +401,6 @@ function tagsEdit() {
   }
 }
 
-function showCustomMode2Popup(mode) {
-  if ($("#customMode2PopupWrapper").hasClass("hidden")) {
-    if (mode == "time") {
-      $("#customMode2Popup .title").text("Test length");
-      $("#customMode2Popup").attr("mode", "time");
-    } else if (mode == "words") {
-      $("#customMode2Popup .title").text("Word amount");
-      $("#customMode2Popup").attr("mode", "words");
-    }
-    $("#customMode2PopupWrapper")
-      .stop(true, true)
-      .css("opacity", 0)
-      .removeClass("hidden")
-      .animate({ opacity: 1 }, 100, (e) => {
-        $("#customMode2Popup input").focus().select();
-      });
-  }
-}
-
-function hideCustomMode2Popup() {
-  if (!$("#customMode2PopupWrapper").hasClass("hidden")) {
-    $("#customMode2PopupWrapper")
-      .stop(true, true)
-      .css("opacity", 1)
-      .animate(
-        {
-          opacity: 0,
-        },
-        100,
-        (e) => {
-          $("#customMode2PopupWrapper").addClass("hidden");
-        }
-      );
-  }
-}
-
-$("#customMode2PopupWrapper").click((e) => {
-  if ($(e.target).attr("id") === "customMode2PopupWrapper") {
-    hideCustomMode2Popup();
-  }
-});
-
-$("#customMode2Popup input").keypress((e) => {
-  if (e.keyCode == 13) {
-    applyMode2Popup();
-  }
-});
-
-$("#customMode2Popup .button").click(() => {
-  applyMode2Popup();
-});
-
-function applyMode2Popup() {
-  let mode = $("#customMode2Popup").attr("mode");
-  let val = parseInt($("#customMode2Popup input").val());
-
-  if (mode == "time") {
-    if (val !== null && !isNaN(val) && val >= 0) {
-      UpdateConfig.setTimeConfig(val);
-      ManualRestart.set();
-      TestLogic.restart();
-      if (val >= 1800) {
-        Notifications.add("Stay safe and take breaks!", 0);
-      } else if (val == 0) {
-        Notifications.add(
-          "Infinite time! Make sure to use Bail Out from the command line to save your result.",
-          0,
-          7
-        );
-      }
-    } else {
-      Notifications.add("Custom time must be at least 1", 0);
-    }
-  } else if (mode == "words") {
-    if (val !== null && !isNaN(val) && val >= 0) {
-      UpdateConfig.setWordCount(val);
-      ManualRestart.set();
-      TestLogic.restart();
-      if (val > 2000) {
-        Notifications.add("Stay safe and take breaks!", 0);
-      } else if (val == 0) {
-        Notifications.add(
-          "Infinite words! Make sure to use Bail Out from the command line to save your result.",
-          0,
-          7
-        );
-      }
-    } else {
-      Notifications.add("Custom word amount must be at least 1", 0);
-    }
-  }
-
-  hideCustomMode2Popup();
-}
-
 $(document).on("click", "#top .logo", (e) => {
   changePage("test");
 });
@@ -503,7 +408,7 @@ $(document).on("click", "#top .logo", (e) => {
 $(document).on("click", "#top .config .wordCount .text-button", (e) => {
   const wrd = $(e.currentTarget).attr("wordCount");
   if (wrd == "custom") {
-    showCustomMode2Popup("words");
+    CustomMode2Popup.show("words");
   } else {
     UpdateConfig.setWordCount(wrd);
     ManualRestart.set();
@@ -514,7 +419,7 @@ $(document).on("click", "#top .config .wordCount .text-button", (e) => {
 $(document).on("click", "#top .config .time .text-button", (e) => {
   let mode = $(e.currentTarget).attr("timeConfig");
   if (mode == "custom") {
-    showCustomMode2Popup("time");
+    CustomMode2Popup.show("time");
   } else {
     UpdateConfig.setTimeConfig(mode);
     ManualRestart.set();
@@ -1811,87 +1716,3 @@ window.addEventListener("keydown", function (e) {
     e.preventDefault();
   }
 });
-
-async function setupChallenge(challengeName) {
-  let list = await Misc.getChallengeList();
-  let challenge = list.filter((c) => c.name === challengeName)[0];
-  let notitext;
-  try {
-    if (challenge === undefined) {
-      Notifications.add("Challenge not found", 0);
-      ManualRestart.set();
-      TestLogic.restart(false, true);
-      setTimeout(() => {
-        $("#top .config").removeClass("hidden");
-        $(".page.pageTest").removeClass("hidden");
-      }, 250);
-      return;
-    }
-    if (challenge.type === "customTime") {
-      UpdateConfig.setTimeConfig(challenge.parameters[0], true);
-      UpdateConfig.setMode("time", true);
-      UpdateConfig.setDifficulty("normal", true);
-      if (challenge.name === "englishMaster") {
-        UpdateConfig.setLanguage("english_10k", true);
-        UpdateConfig.setNumbers(true, true);
-        UpdateConfig.setPunctuation(true, true);
-      }
-    }
-    if (challenge.type === "customWords") {
-      UpdateConfig.setWordCount(challenge.parameters[0], true);
-      UpdateConfig.setMode("words", true);
-      UpdateConfig.setDifficulty("normal", true);
-    } else if (challenge.type === "customText") {
-      CustomText.setText(challenge.parameters[0].split(" "));
-      CustomText.setIsWordRandom(challenge.parameters[1]);
-      CustomText.setWord(parseInt(challenge.parameters[2]));
-      UpdateConfig.setMode("custom", true);
-      UpdateConfig.setDifficulty("normal", true);
-    } else if (challenge.type === "script") {
-      let scriptdata = await fetch("/challenges/" + challenge.parameters[0]);
-      scriptdata = await scriptdata.text();
-      let text = scriptdata.trim();
-      text = text.replace(/[\n\r\t ]/gm, " ");
-      text = text.replace(/ +/gm, " ");
-      CustomText.setText(text.split(" "));
-      CustomText.setIsWordRandom(false);
-      UpdateConfig.setMode("custom", true);
-      UpdateConfig.setDifficulty("normal", true);
-      if (challenge.parameters[1] != null) {
-        UpdateConfig.setTheme(challenge.parameters[1]);
-      }
-      if (challenge.parameters[2] != null) {
-        Funbox.activate(challenge.parameters[2]);
-      }
-    } else if (challenge.type === "accuracy") {
-      UpdateConfig.setTimeConfig(0, true);
-      UpdateConfig.setMode("time", true);
-      UpdateConfig.setDifficulty("master", true);
-    } else if (challenge.type === "funbox") {
-      Funbox.activate(challenge.parameters[0]);
-      UpdateConfig.setDifficulty("normal", true);
-      if (challenge.parameters[1] === "words") {
-        UpdateConfig.setWordCount(challenge.parameters[2], true);
-      } else if (challenge.parameters[1] === "time") {
-        UpdateConfig.setTimeConfig(challenge.parameters[2], true);
-      }
-      UpdateConfig.setMode(challenge.parameters[1], true);
-      if (challenge.parameters[3] !== undefined) {
-        UpdateConfig.setDifficulty(challenge.parameters[3], true);
-      }
-    }
-    ManualRestart.set();
-    TestLogic.restart(false, true);
-    notitext = challenge.message;
-    $("#top .config").removeClass("hidden");
-    $(".page.pageTest").removeClass("hidden");
-
-    if (notitext === undefined) {
-      Notifications.add(`Challenge '${challengeName}' loaded.`, 0);
-    } else {
-      Notifications.add("Challenge loaded. " + notitext, 0);
-    }
-  } catch (e) {
-    Notifications.add("Something went wrong: " + e, -1);
-  }
-}
